+++ lib/src/banner/controller.dart
@@ -160,6 +160,11 @@ class BannerAdController extends LoadShowAd<BannerAdEvent>
   /// For more info, [read the documentation](https://github.com/bdlukaa/native_admob_flutter/wiki/Using-the-controller-and-listening-to-banner-events#listening-to-events)
   Stream<Map<BannerAdEvent, dynamic>> get onEvent => super.onEvent;
 
+  bool _loaded = false;
+
+  /// Returns true if the ad was successfully loaded and is ready to be rendered.
+  bool get isLoaded => _loaded;
+
   /// Creates a new native ad controller
   BannerAdController() : super();
 
@@ -192,14 +197,17 @@ class BannerAdController extends LoadShowAd<BannerAdEvent>
     if (isDisposed) return;
     switch (call.method) {
       case 'loading':
+        _loaded = false;
         onEventController.add({BannerAdEvent.loading: null});
         break;
       case 'onAdFailedToLoad':
+        _loaded = false;
         onEventController.add({
           BannerAdEvent.loadFailed: AdError.fromJson(call.arguments),
         });
         break;
       case 'onAdLoaded':
+        _loaded = true;
         onEventController.add({BannerAdEvent.loaded: call.arguments});
         break;
       case 'onAdImpression':
@@ -213,9 +221,14 @@ class BannerAdController extends LoadShowAd<BannerAdEvent>
   /// Load the ad. The ad needs to be loaded to be rendered.
   ///
   /// For more info, read the [documentation](https://github.com/bdlukaa/native_admob_flutter/wiki/Using-the-controller-and-listening-to-banner-events#reloading-the-ad)
-  Future<bool> load() {
+  Future<bool> load({
+    /// Force to load an ad even if another is already avaiable
+    bool force = false,
+  }) async {
     ensureAdNotDisposed();
     assertMobileAdsIsInitialized();
-    return channel.invokeMethod<bool>('loadAd');
+    if (!debugCheckAdWillReload(isLoaded, force)) return false;
+    _loaded = await channel.invokeMethod<bool>('loadAd');
+    return isLoaded;
   }
 }