+++ lib/src/banner/options.dart
@@ -1,0 +1,10 @@
+class BannerAdOptions {
+  final bool reloadWhenOrientationChanges;
+  final bool reloadWhenSizeChanges;
+
+  const BannerAdOptions({
+    this.reloadWhenOrientationChanges = true,
+    this.reloadWhenSizeChanges = true,
+  })  : assert(reloadWhenOrientationChanges != null),
+        assert(reloadWhenSizeChanges != null);
+}