+++ lib/src/rewarded/rewarded.dart
@@ -15,40 +15,93 @@ import '../utils.dart';
 ///   - showed (When the ad is opened)
 ///   - failedToShow (When it failed to show the ad)
 ///   - showFailed (When the ad failed to show)
-///   = earnedReward (When the user earns a reward)
+///   - earnedReward (When the user earns a reward)
 ///
 /// For more info on rewarded ad events, read the [documentation](https://github.com/bdlukaa/native_admob_flutter/wiki/Creating-a-rewarded-ad#listening-to-events)
 enum RewardedAdEvent {
-  /// Called when an ad request failed.
+  /// Called when the ad failed to load. This event is usually thrown by `ad.load()`
   ///
   /// **Warning**: Attempting to load a new ad when the load fail
   /// is strongly discouraged. If you must load an ad when it fails,
   /// limit ad load retries to avoid continuous failed ad requests in
   /// situations such as limited network connectivity.
+  ///
+  /// You can check the reason like this
+  /// ```dart
+  /// rewardedAd.onEvent.listen((e) {
+  ///   final event = e.keys.first;
+  ///     switch (event) {
+  ///       case RewardedAdEvent.loadFailed:
+  ///         final error = e.values.first;
+  ///         print('load failed ${error.code}');
+  ///         break;
+  ///       default:
+  ///         break;
+  ///     }
+  /// });
+  /// ```
+  ///
+  /// For more info, read the [documentation](https://github.com/bdlukaa/native_admob_flutter/wiki/Creating-a-rewarded-ad#load-the-ad)
   loadFailed,
 
-  /// Called when an ad is loaded.
+  /// Called when an ad is loaded. This event is usually thrown by `ad.load()`
   ///
   /// You can also check if the ad is loaded by calling `ad.isLoaded`.
   /// If it's not loaded, you can load a new ad by calling `ad.load()`
+  ///
+  /// For more info, read the [documentation](https://github.com/bdlukaa/native_admob_flutter/wiki/Creating-a-rewarded-ad#load-the-ad)
   loaded,
 
-  /// Called when the ad starts loading.
-  /// This event is usually thrown by `ad.load()`
+  /// Called when the ad starts loading. This event is usually thrown by `ad.load()`
+  ///
+  /// For more info, read the [documentation](https://github.com/bdlukaa/native_admob_flutter/wiki/Creating-a-rewarded-ad#load-the-ad)
   loading,
 
-  /// Called when the ad showed.
-  /// This event is usually thrown by `ad.show()`
+  /// Called when the ad showed. This event is usually thrown by `ad.show()`
+  ///
+  /// For more info, read the [documentation](https://github.com/bdlukaa/native_admob_flutter/wiki/Creating-a-rewarded-ad#show-the-ad)
   showed,
 
   /// Called when the ad failed to show.
   /// This event is usually thrown by `ad.show()`
+  ///
+  /// You can check the reason like this
+  /// ```dart
+  /// rewardedAd.onEvent.listen((e) {
+  ///   final event = e.keys.first;
+  ///     switch (event) {
+  ///       case RewardedAdEvent.showFailed:
+  ///         final error = e.values.first;
+  ///         print('show failed ${error.code}');
+  ///         break;
+  ///       default:
+  ///         break;
+  ///     }
+  /// });
+  /// ```
+  ///
+  /// For more info, read the [documentation](https://github.com/bdlukaa/native_admob_flutter/wiki/Creating-a-rewarded-ad#show-the-ad)
   showFailed,
 
-  /// Called when the ad closes
+  /// Called when the ad is closed
   closed,
 
-  /// Called when the user earns an reward
+  /// Called when the user earns an reward. You can check
+  /// the reward item like this:
+  ///
+  /// ```dart
+  /// rewardedAd.onEvent.listen((e) {
+  ///   final event = e.keys.first;
+  ///     switch (event) {
+  ///       case RewardedAdEvent.earnedReward:
+  ///         final reward = e.values.first;
+  ///         print('earned reward: $reward');
+  ///         break;
+  ///       default:
+  ///         break;
+  ///     }
+  /// });
+  /// ```
   earnedReward,
 }
 
@@ -97,15 +150,15 @@ class RewardedAd extends LoadShowAd<RewardedAdEvent> {
   ///       print('loaded');
   ///       break;
   ///     case RewardedAdEvent.loadFailed:
-  ///       final errorCode = e.values.first;
-  ///       print('load failed $errorCode');
+  ///       final error = e.values.first;
+  ///       print('load failed $error');
   ///       break;
   ///     case RewardedAdEvent.showed:
   ///       print('ad showed');
   ///       break;
   ///     case RewardedAdEvent.showFailed:
-  ///       final errorCode = e.values.first;
-  ///       print('show failed $errorCode');
+  ///       final error = e.values.first;
+  ///       print('show failed $error');
   ///       break;
   ///     case RewardedAdEvent.closed:
   ///       print('ad closed');
@@ -207,13 +260,17 @@ class RewardedAd extends LoadShowAd<RewardedAdEvent> {
   /// ```
   ///
   /// For more info, [read the documentation](https://github.com/bdlukaa/native_admob_flutter/wiki/Creating-a-rewarded-ad#load-the-ad)
-  Future<bool> load({String unitId}) async {
+  Future<bool> load({
+    /// The ad unit id. If null, uses [MobileAds.rewardedAdUnitId]
+    String unitId,
+
+    /// Force to load an ad even if another is already avaiable
+    bool force = false,
+  }) async {
+    assert(force != null);
     ensureAdNotDisposed();
     assertMobileAdsIsInitialized();
-    if (isLoaded) {
-      print('An ad is already avaiable, no need to load another');
-      return false;
-    }
+    if (!debugCheckAdWillReload(isLoaded, force)) return false;
     _loaded = await channel.invokeMethod<bool>('loadAd', {
       'unitId': unitId ??
           MobileAds.rewardedAdUnitId ??